+++ frontend/src/features/halls/context/hall-context.tsx
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { HallContextType } from "@/types";
+
+export const HallContext = createContext<HallContextType | undefined>(undefined);
